Remove embedded certificates
